ADD: upload images FINISH: code review

+++ server/controllers/likeController.js
@@ -8,17 +8,21 @@ export const likePost = async (req, res) => {
 
     const existingLike = await Like.findOne({ post: postId, author: userId });
     if (existingLike) {
+      const post = await Post.findById(postId);
+      if (!post) return res.status(404).json('Post not found');
+
       await Like.findByIdAndDelete(existingLike.id);
 
-      const post = await Post.findById(postId);
       post.likes--;
       await post.save();
 
       res.status(200).json(`You unliked the post!`);
     } else {
+      const post = await Post.findById(postId);
+      if (!post) return res.status(404).json('Post not found');
+
       const like = await Like.create({ post: postId, author: userId });
 
-      const post = await Post.findById(postId);
       post.likes++;
       await post.save();
